vcs: Move out all `WorkingTree`s to separate files
This increases readability and implicitly gives the formerly anonymous
classes names so they can later be referred to. Slightly reorder functions
while at it, so overridden functions go first and private functions go
last.
