+++ downloader/src/main/kotlin/vcs/Cvs.kt
@@ -20,7 +20,6 @@
 package org.ossreviewtoolkit.downloader.vcs
 
 import java.io.File
-import java.security.MessageDigest
 import java.util.regex.Pattern
 
 import org.ossreviewtoolkit.downloader.VersionControlSystem
@@ -28,10 +27,6 @@ import org.ossreviewtoolkit.downloader.WorkingTree
 import org.ossreviewtoolkit.model.VcsInfo
 import org.ossreviewtoolkit.model.VcsType
 import org.ossreviewtoolkit.utils.common.CommandLineTool
-import org.ossreviewtoolkit.utils.common.ProcessCapture
-import org.ossreviewtoolkit.utils.common.encodeHex
-import org.ossreviewtoolkit.utils.common.safeDeleteRecursively
-import org.ossreviewtoolkit.utils.common.searchUpwardsForSubdirectory
 
 typealias CvsFileRevisions = List<Pair<String, String>>
 
@@ -58,120 +53,7 @@ class Cvs : VersionControlSystem() {
 
     override fun getDefaultBranchName(url: String): String? = null
 
-    override fun getWorkingTree(vcsDirectory: File) =
-        object : WorkingTree(vcsDirectory, type) {
-            private val cvsDirectory = workingDir.resolve("CVS")
-
-            override fun isValid(): Boolean {
-                if (!workingDir.isDirectory) {
-                    return false
-                }
-
-                return ProcessCapture(workingDir, "cvs", "status", "-l").isSuccess
-            }
-
-            override fun isShallow() = false
-
-            override fun getRemoteUrl() = cvsDirectory.resolve("Root").useLines { it.first() }
-
-            override fun getRevision() =
-                // CVS does not have the concept of a global revision, but each file has its own revision. As
-                // we just use the revision to uniquely identify the state of a working tree, simply create an
-                // artificial single revision based on the revisions of all files.
-                getFileRevisionsHash(getFileRevisions())
-
-            private fun getFileRevisions(): CvsFileRevisions {
-                val cvsLog = CvsCommand.run(workingDir, "log", "-h")
-
-                var currentWorkingFile = ""
-                return cvsLog.stdout.lines().mapNotNull { line ->
-                    var value = line.removePrefix("Working file: ")
-                    if (value.length < line.length) {
-                        currentWorkingFile = value
-                    } else {
-                        value = line.removePrefix("head: ")
-                        if (value.length < line.length) {
-                            if (currentWorkingFile.isNotBlank() && value.isNotBlank()) {
-                                return@mapNotNull Pair(currentWorkingFile, value)
-                            }
-                        }
-                    }
-
-                    null
-                }.sortedBy { it.first }
-            }
-
-            private fun getFileRevisionsHash(fileRevisions: CvsFileRevisions) =
-                fileRevisions.fold(MessageDigest.getInstance("SHA-1")) { digest, (file, revision) ->
-                    digest.apply {
-                        update(file.toByteArray())
-                        update(revision.toByteArray())
-                    }
-                }.digest().encodeHex()
-
-            override fun getRootPath(): File {
-                val rootDir = workingDir.searchUpwardsForSubdirectory("CVS")
-                return rootDir ?: workingDir
-            }
-
-            private fun listSymbolicNames(): Map<String, String> =
-                try {
-                    // Create all working tree directories in order to be able to query the log.
-                    CvsCommand.run(workingDir, "update", "-d")
-
-                    val cvsLog = CvsCommand.run(workingDir, "log", "-h")
-                    var tagsSectionStarted = false
-
-                    cvsLog.stdout.lines().mapNotNull { line ->
-                        if (tagsSectionStarted) {
-                            if (line.startsWith('\t')) {
-                                line.split(':', limit = 2).let {
-                                    Pair(it.first().trim(), it.last().trim())
-                                }
-                            } else {
-                                tagsSectionStarted = false
-                                null
-                            }
-                        } else {
-                            if (line == "symbolic names:") {
-                                tagsSectionStarted = true
-                            }
-                            null
-                        }
-                    }.toMap().toSortedMap()
-                } finally {
-                    // Clean the temporarily updated working tree again.
-                    workingDir.walk().maxDepth(1).forEach {
-                        if (it.isDirectory) {
-                            if (it != workingDir && it.name != "CVS") it.safeDeleteRecursively()
-                        } else {
-                            it.delete()
-                        }
-                    }
-                }
-
-            private fun isBranchVersion(version: String): Boolean {
-                // See http://cvsgui.sourceforge.net/howto/cvsdoc/cvs_5.html#SEC59.
-                val decimals = version.split('.')
-
-                // "Externally, branch numbers consist of an odd number of dot-separated decimal
-                // integers."
-                return decimals.size % 2 != 0 ||
-                        // "That is not the whole truth, however. For efficiency reasons CVS sometimes inserts
-                        // an extra 0 in the second rightmost position."
-                        decimals.dropLast(1).last() == "0"
-            }
-
-            override fun listRemoteBranches() =
-                listSymbolicNames().mapNotNull { (name, version) ->
-                    name.takeIf { isBranchVersion(version) }
-                }
-
-            override fun listRemoteTags() =
-                listSymbolicNames().mapNotNull { (name, version) ->
-                    name.takeUnless { isBranchVersion(version) }
-                }
-        }
+    override fun getWorkingTree(vcsDirectory: File) = CvsWorkingTree(vcsDirectory, type)
 
     override fun isApplicableUrlInternal(vcsUrl: String) = vcsUrl.matches(":(ext|pserver):[^@]+@.+".toRegex())
